optimisations to the chess validation logic, not functional yet

+++ src/main/java/com/sah/game/dtos/MoveCoords.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class MoveCoords {
-    private short fromRow, fromCol, targetRow, targetCol;
+    private int fromRow, fromCol, targetRow, targetCol;
 }
